Deprecate pixel_data_handlers.util
