+++ doc/reference/pixels.decoders.rst
@@ -3,6 +3,8 @@
 Pixel Data Decoders (:mod:`pydicom.pixels.decoders`)
 ====================================================
 
+.. module:: pydicom.pixels.decoders
+
 :class:`~pydicom.pixels.decoders.base.Decoder` class instances for decoding
 native and compressed (7FE0,0010) *Pixel Data*, (7FE0,0008) *Float Pixel Data* and
 (7FE0,0009) *Double Float Pixel Data*
