+++ src/pydicom/pixel_data_handlers/pillow_handler.py
@@ -31,7 +31,7 @@ except ImportError:
 from pydicom import config
 from pydicom.encaps import generate_frames
 from pydicom.misc import warn_and_log
-from pydicom.pixel_data_handlers.util import (
+from pydicom.pixels.utils import (
     pixel_dtype,
     get_j2k_parameters,
     get_nr_frames,
