+++ src/pydicom/pixel_data_handlers/pylibjpeg_handler.py
@@ -99,7 +99,7 @@ except ImportError:
 
 from pydicom import config
 from pydicom.encaps import generate_frames as frame_generator
-from pydicom.pixel_data_handlers.util import (
+from pydicom.pixels.utils import (
     pixel_dtype,
     get_expected_length,
     reshape_pixel_array,
