+++ src/pydicom/pixel_data_handlers/rle_handler.py
@@ -48,7 +48,7 @@ except ImportError:
 
 from pydicom.encaps import generate_frames
 from pydicom.misc import warn_and_log
-from pydicom.pixel_data_handlers.util import pixel_dtype, get_nr_frames
+from pydicom.pixels.utils import pixel_dtype, get_nr_frames
 import pydicom.uid
 
 if TYPE_CHECKING:  # pragma: no cover
