+++ src/pydicom/pixels/decoders/__init__.py
@@ -1,0 +1,18 @@
+from pydicom.pixels.decoders.base import (
+    ImplicitVRLittleEndianDecoder,
+    ExplicitVRLittleEndianDecoder,
+    ExplicitVRBigEndianDecoder,
+    DeflatedExplicitVRLittleEndianDecoder,
+    JPEGBaseline8BitDecoder,
+    JPEGExtended12BitDecoder,
+    JPEGLosslessDecoder,
+    JPEGLosslessSV1Decoder,
+    JPEGLSLosslessDecoder,
+    JPEGLSNearLosslessDecoder,
+    JPEG2000LosslessDecoder,
+    JPEG2000Decoder,
+    HTJ2KLosslessDecoder,
+    HTJ2KLosslessRPCLDecoder,
+    HTJ2KDecoder,
+    RLELosslessDecoder,
+)
