+++ src/pydicom/pixels/encoders/__init__.py
@@ -1,0 +1,5 @@
+from pydicom.pixels.encoders.base import (
+    RLELosslessEncoder,
+    JPEGLSLosslessEncoder,
+    JPEGLSNearLosslessEncoder,
+)
